requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==3.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.108.0
fastapi-oauth2==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
libgravatar==1.0.4
Mako==1.3.0
MarkupSafe==2.1.3
oauthlib==3.2.2
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydentic==0.0.1.dev3
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
python-stdnum==1.19
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
social-auth-core==4.5.3
SQLAlchemy==2.0.24
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.25.0
